fix: unit tests
moved eigenda_client test to disperser_client test - it was in the wrong spot
